requirements.txt
Flask>=2.0.1
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
requests>=2.26.0
cachecontrol>=0.12.6
pathlib>=1.0.1
gunicorn==20.1.0
